I answered the questions.
But the markdown language isn't working, at all. The only thing that does something is the "*" which does  not indent the line with a dot but italicizes all the words... I'm going to turn it in like this and figure this out in the next lab. I'm familiar with markdown languages having worked with HTML. I just don't know why I'm having such technical difficulties in this first week. Growing Pains? ;)